Organized everything better
